got the thankyou page and thank you page editable working

+++ Backend/src/main/java/com/smartOrder/restaurant_managment_app/Models/Users.java
@@ -117,13 +117,13 @@ public void setResetTokenExpiry(LocalDateTime resetTokenExpiry) {
   }
   
   @Column(name = "restaurant_id", nullable = false)
-  private Integer restaurantId; 
+  private Long restaurantId; 
   
   
-  public Integer getRestaurantId() {
+  public Long getRestaurantId() {
     return restaurantId;
   }
-  public void setRestaurantId(Integer restaurantId) {
+  public void setRestaurantId(Long restaurantId) {
     this.restaurantId = restaurantId;
   }
