+++ Backend/src/main/java/com/smartOrder/restaurant_managment_app/repository/UserRepo.java
@@ -14,7 +14,7 @@ public interface UserRepo extends JpaRepository<Users, Integer> {
   boolean existsByUsername(String username);
   List<Users> findByRestaurantId(Integer restaurantId);
   List<Users> findByRestaurantIdAndRole(Integer restaurantId, Users.Role role);
-  boolean existsByRestaurantIdAndRole(Integer restaurantId, Users.Role role);
+  boolean existsByRestaurantIdAndRole(Long restaurantId, Users.Role role);
   boolean existsById(Long id);
   void deleteById(Long id);
   Users findByResetToken(String resetToken);
