+++ Backend/src/main/java/com/smartOrder/restaurant_managment_app/services/UserService.java
@@ -94,7 +94,7 @@ public class UserService {
 
     public Users createUser(Users user) {
       // Verify restaurant exists
-      if (!restaurantRepository.existsById(user.getRestaurantId())) {
+      if (!restaurantRepository.existsById(user.getRestaurantId().intValue())) {
           throw new RuntimeException("Restaurant not found");
       }
